+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClient.java
@@ -2,12 +2,15 @@ package com.getindata.connectors.http.internal.table.lookup;
 
 import java.io.IOException;
 import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.net.http.HttpResponse.BodyHandlers;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -17,16 +20,20 @@ import com.fasterxml.jackson.databind.node.NullNode;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
+import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.util.StringUtils;
 
 import com.getindata.connectors.http.HttpPostRequestCallback;
+import com.getindata.connectors.http.internal.HeaderPreprocessor;
 import com.getindata.connectors.http.internal.PollingClient;
 import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
 import com.getindata.connectors.http.internal.status.ComposeHttpStatusCodeChecker;
 import com.getindata.connectors.http.internal.status.ComposeHttpStatusCodeChecker.ComposeHttpStatusCodeCheckerConfig;
 import com.getindata.connectors.http.internal.status.HttpStatusCodeChecker;
+import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
 import static com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants.RESULT_TYPE;
+import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.SOURCE_LOOKUP_OIDC_AUTH_TOKEN_REQUEST;
 
 /**
  * An implementation of {@link PollingClient} that uses Java 11's {@link HttpClient}.
@@ -96,10 +103,57 @@ public class JavaNetHttpPollingClient implements PollingClient<RowData> {
 
         HttpLookupSourceRequestEntry request = requestFactory.buildLookupRequest(lookupData);
         HttpResponse<String> response = httpClient.send(
-            request.getHttpRequest(),
-            BodyHandlers.ofString()
-        );
+            updateHttpRequestIfRequired(request,
+                    HttpHeaderUtils.createOIDCHeaderPreprocessor(options.getReadableConfig())),
+            BodyHandlers.ofString());
         return processHttpResponse(response, request);
+    }
+
+    /**
+     * If using OIDC, update the http request using the oidc header pre processor to supply the
+     * authentication header, with a short lived bearer token.
+     * @param request http reauest to amend
+     * @param oidcHeaderPreProcessor OIDC header pre processor
+     * @return http request, which for OIDC will have the bearer token as the authentication header
+     */
+    protected HttpRequest updateHttpRequestIfRequired(HttpLookupSourceRequestEntry request,
+                                                      HeaderPreprocessor oidcHeaderPreProcessor) {
+        // We need to check the config and if required amend the value of the
+        // authentication header to the short lived bearer token
+        HttpRequest httpRequest = request.getHttpRequest();
+        ReadableConfig readableConfig = options.getReadableConfig();
+        if (oidcHeaderPreProcessor != null) {
+            HttpRequest.Builder builder = HttpRequest.newBuilder()
+                    .uri(httpRequest.uri());
+            if (httpRequest.timeout().isPresent()) {
+                builder.timeout(httpRequest.timeout().get());
+            }
+            if (httpRequest.method().endsWith("GET")) {
+                builder.GET();
+            } else {
+                builder.method(httpRequest.method(), httpRequest.bodyPublisher().get());
+            }
+            Map<String, String> headerMap = new HashMap<>();
+            if (httpRequest.headers() != null && !httpRequest.headers().map().isEmpty()) {
+                for (Map.Entry<String, List<String>> header
+                        :httpRequest.headers().map().entrySet()) {
+                    List<String> values =  header.getValue();
+                    if (values.size() == 1) {
+                        headerMap.put(header.getKey(), header.getValue().get(0));
+                    }
+                    // the existing design does not handle multiple values for headers
+                }
+            }
+            Optional<String> oidcTokenRequest = readableConfig
+                    .getOptional(SOURCE_LOOKUP_OIDC_AUTH_TOKEN_REQUEST);
+            String bearerToken = oidcHeaderPreProcessor.preprocessValueForHeader(
+                    HttpHeaderUtils.AUTHORIZATION,  oidcTokenRequest.get());
+            headerMap.put(HttpHeaderUtils.AUTHORIZATION, bearerToken);
+            String[] headerAndValueArray = HttpHeaderUtils.toHeaderAndValueArray(headerMap);
+            builder.headers(headerAndValueArray);
+            httpRequest = builder.build();
+        }
+        return httpRequest;
     }
 
     private Collection<RowData> processHttpResponse(